requirements.txt
# Web Framework
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0

# Task Queue & Async Processing
celery[redis]==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.9

# Monitoring
flower==2.0.1

# Configuration
python-dotenv==1.0.0

# Utilities
urllib3==2.1.0
